# public/swagger.json
{"swagger":"2.0","info":{"version":"1.1.0.0","title":"Api movies"},"host":"localhost:10010","basePath":"/","schemes":["http","https"],"consumes":["application/json"],"produces":["application/json"],"paths":{"/movie":{"x-swagger-router-controller":"movie","get":{"operationId":"getAll","description":"get the movies list","responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GetMoviesListResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}},"post":{"operationId":"save","description":"add a new movie to the list","parameters":[{"name":"title","description":"Movie properties","in":"body","required":true,"schema":{"$ref":"#/definitions/Movie"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GeneralResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}},"/movie/{id}":{"x-swagger-router-controller":"movie","get":{"operationId":"getOne","description":"get a movie","parameters":[{"name":"id","type":"string","in":"path","required":true}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GetMovieResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}},"put":{"operationId":"update","description":"update a movie","parameters":[{"name":"id","description":"Movie id","type":"string","in":"path","required":true},{"name":"title","description":"Movie properties","in":"body","required":true,"schema":{"$ref":"#/definitions/Movie"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GeneralResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}},"patch":{"operationId":"patch","description":"patch a movie","parameters":[{"name":"id","description":"Movie id","type":"string","in":"path","required":true},{"name":"title","description":"Movie properties","in":"body","required":true,"schema":{"$ref":"#/definitions/Movie"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GeneralResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}},"delete":{"operationId":"delMovie","description":"delete a movie","parameters":[{"name":"id","description":"Movie id","type":"string","in":"path","required":true}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/GeneralResponse"}},"default":{"description":"Error","schema":{"$ref":"#/definitions/ErrorResponse"}}}}}},"definitions":{"GetMoviesListResponse":{"required":["movies"],"properties":{"movies":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"year":{"type":"number"}}}}}},"GeneralResponse":{"type":"object","properties":{"success":{"type":"number","description":"returns 1 if successful"},"description":{"type":"string","description":"a short comment"}},"required":["success","description"]},"ErrorResponse":{"required":["message"],"properties":{"message":{"type":"string"}}},"GetMovieResponse":{"required":["id","title","year"],"properties":{"id":{"type":"string"},"title":{"type":"string"},"year":{"type":"number"}}},"Movie":{"type":"object","properties":{"title":{"type":"string","description":"task object name"},"year":{"type":"number","description":"task description"}},"required":["title","year"]}}}